submit solution for hw2 without tests and very slow :(

+++ MJT/hw2/src/bg/sofia/uni/fmi/mjt/goodreads/recommender/similaritycalculator/genres/GenresOverlapSimilarityCalculator.java
@@ -11,10 +11,10 @@ public class GenresOverlapSimilarityCalculator implements SimilarityCalculator {
             throw new IllegalArgumentException("Cannot calculate similarity between null objects");
         }
 
-        return (double) first.genres().parallelStream()
+        return first.genres().parallelStream()
             .filter(genre -> second.genres().contains(genre))
             .toList()
-            .size() / Integer.min(first.genres().size(), second.genres().size());
+            .size() / (double) Integer.min(first.genres().size(), second.genres().size());
     }
 
 }
